Split mocking and persistent_storage
