+++ tests/integration/test_github_readonly.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 from ogr import GithubService
-from ogr.services.mock.mock_core import PersistentObjectStorage
+from ogr.persistent_storage import PersistentObjectStorage
 
 DATA_DIR = "test_data"
 PERSISTENT_DATA_PREFIX = os.path.join(
