+++ tests/integration/test_pagure.py
@@ -4,7 +4,7 @@ import unittest
 from ogr import PagureService
 from ogr.abstract import PRStatus, IssueStatus
 from ogr.exceptions import PagureAPIException
-from ogr.services.mock.mock_core import PersistentObjectStorage
+from ogr.persistent_storage import PersistentObjectStorage
 
 DATA_DIR = "test_data"
 PERSISTENT_DATA_PREFIX = os.path.join(
